Refactor Calendar View into Multiple Widgets
