+++ unischedule/lib/presentation/calendar_page/widgets/calendar_background.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class CalendarAppBackground extends StatelessWidget {
+  const CalendarAppBackground({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Opacity(
+      opacity: 0.7,
+      child: Container(
+        decoration: const BoxDecoration(
+          image: DecorationImage(
+            image: AssetImage('assets/images/bg.jpg'),
+            fit: BoxFit.cover,
+          ),
+        ),
+      ),
+    );
+  }
+}
